+++ monkey/monkey_island/cc/resources/configuration_export.py
@@ -4,8 +4,10 @@ import flask_restful
 from flask import request
 
 from monkey_island.cc.resources.auth.auth import jwt_required
+from monkey_island.cc.server_utils.encryption.password_based_encryption import (
+    PasswordBasedEncryptor,
+)
 from monkey_island.cc.services.config import ConfigService
-from monkey_island.cc.services.utils.password_encryption import PasswordBasedEncryptor
 
 
 class ConfigurationExport(flask_restful.Resource):
