+++ monkey/monkey_island/cc/server_utils/encryption/data_store_encryptor.py
@@ -4,8 +4,8 @@ import os
 # is maintained.
 from Crypto import Random  # noqa: DUO133  # nosec: B413
 
+from monkey_island.cc.server_utils.encryption.key_based_encryptor import KeyBasedEncryptor
 from monkey_island.cc.server_utils.file_utils import open_new_securely_permissioned_file
-from monkey_island.cc.services.utils.key_encryption import KeyBasedEncryptor
 
 _encryptor = None
 
@@ -22,6 +22,8 @@ class DataStoreEncryptor:
         else:
             self._init_key(password_file)
 
+        self._key_base_encryptor = KeyBasedEncryptor(self._cipher_key)
+
     def _init_key(self, password_file_path: str):
         self._cipher_key = Random.new().read(self._BLOCK_SIZE)
         with open_new_securely_permissioned_file(password_file_path, "wb") as f:
@@ -32,12 +34,10 @@ class DataStoreEncryptor:
             self._cipher_key = f.read()
 
     def enc(self, message: str):
-        key_encryptor = KeyBasedEncryptor(self._cipher_key)
-        return key_encryptor.encrypt(message)
+        return self._key_base_encryptor.encrypt(message)
 
     def dec(self, enc_message: str):
-        key_encryptor = KeyBasedEncryptor(self._cipher_key)
-        return key_encryptor.decrypt(enc_message)
+        return self._key_base_encryptor.decrypt(enc_message)
 
 
 def initialize_encryptor(password_file_dir):
