+++ monkey/monkey_island/cc/server_utils/encryption/i_encryptor.py
