+++ monkey/monkey_island/cc/server_utils/encryption/password_based_encryption.py
@@ -4,7 +4,7 @@ import logging
 
 import pyAesCrypt
 
-from monkey_island.cc.services.utils.i_encryptor import IEncryptor
+from monkey_island.cc.server_utils.encryption.i_encryptor import IEncryptor
 
 logger = logging.getLogger(__name__)
