+++ monkey/monkey_island/cc/services/config.py
@@ -19,7 +19,7 @@ from common.config_value_paths import (
     USER_LIST_PATH,
 )
 from monkey_island.cc.database import mongo
-from monkey_island.cc.server_utils.key_encryptor import get_encryptor
+from monkey_island.cc.server_utils.encryption.data_store_encryptor import get_encryptor
 from monkey_island.cc.services.config_manipulator import update_config_per_mode
 from monkey_island.cc.services.config_schema.config_schema import SCHEMA
 from monkey_island.cc.services.mode.island_mode_service import ModeNotSetError, get_mode
