+++ monkey/monkey_island/cc/services/telemetry/processing/exploit.py
@@ -3,7 +3,7 @@ import copy
 import dateutil
 
 from monkey_island.cc.models import Monkey
-from monkey_island.cc.server_utils.key_encryptor import get_encryptor
+from monkey_island.cc.server_utils.encryption.data_store_encryptor import get_encryptor
 from monkey_island.cc.services.config import ConfigService
 from monkey_island.cc.services.edge.displayed_edge import EdgeService
 from monkey_island.cc.services.node import NodeService
