+++ monkey/monkey_island/cc/services/telemetry/processing/system_info.py
@@ -1,6 +1,6 @@
 import logging
 
-from monkey_island.cc.server_utils.key_encryptor import get_encryptor
+from monkey_island.cc.server_utils.encryption.data_store_encryptor import get_encryptor
 from monkey_island.cc.services.config import ConfigService
 from monkey_island.cc.services.node import NodeService
 from monkey_island.cc.services.telemetry.processing.system_info_collectors.system_info_telemetry_dispatcher import (  # noqa: E501
