+++ monkey/tests/unit_tests/monkey_island/cc/resources/test_configuration_import.py
@@ -1,12 +1,16 @@
 import pytest
+from tests.unit_tests.monkey_island.cc.server_utils.encryption.test_password_based_encryption import (  # noqa: E501
+    PASSWORD,
+)
 from tests.unit_tests.monkey_island.cc.services.utils.ciphertexts_for_encryption_test import (
     MALFORMED_CIPHER_TEXT_CORRUPTED,
 )
-from tests.unit_tests.monkey_island.cc.services.utils.test_encryption import PASSWORD
 
 from common.utils.exceptions import InvalidConfigurationError
 from monkey_island.cc.resources.configuration_import import ConfigurationImport
-from monkey_island.cc.services.utils.password_encryption import PasswordBasedEncryptor
+from monkey_island.cc.server_utils.encryption.password_based_encryption import (
+    PasswordBasedEncryptor,
+)
 
 
 def test_is_config_encrypted__json(monkey_config_json):
